rename AdminComponent to SubscriptionsComponent and move to Businesses namespace
