+++ app/views/businesses/show.html.erb
@@ -41,4 +41,4 @@
   </div>
 </div>
 
-<%= render BusinessAdminComponent.new(@business, user: current_user) %>
+<%= render Businesses::SubscriptionsComponent.new(@business, user: current_user) %>
