+++ config/locales/en.yml
@@ -105,10 +105,6 @@ en:
       cancel: Cancel
       message: Are you sure you want to flag this conversation and block the sender? This will hide the conversation and they won't be able to contact you again.
       title: Block sender
-  business_admin_component:
-    plan: Plan
-    status: Status
-    title: Business subscription
   business_primary_action_component:
     edit: Edit
   businesses:
@@ -139,6 +135,10 @@ en:
       your_business_profile: Your business profile
     show:
       bio: Bio
+    subscriptions_component:
+      plan: Plan
+      status: Status
+      title: Business subscription
     update:
       updated: Your business was updated!
   cold_messages:
